Split theme settings persistence
