+++ doc/architecture/refactor-roadmap.md
@@ -47,6 +47,7 @@ The next refactor should not reduce layering for its own sake. It should turn pa
 - Theme built-in descriptors and token presets live in `ui_theme_builtin.c`, with private theme-system declarations in `ui_theme_internal.h`.
 - External theme registry state, file/directory reload, reload errors, and directory signatures live in `ui_theme_external.c`.
 - Theme JSON extraction, color parsing, path-derived IDs, override application, and token clamping live in `ui_theme_parse.c`.
+- Selected theme ID settings load/save lives in `ui_theme_settings.c`.
 - `render_system_draw()` consumes `RenderSceneDesc`.
 - This refactor round is complete for command routing, workspace API coupling, view-model construction, and low-risk UI composition cleanup.
