+++ src/ui/ui_theme.c
@@ -135,102 +135,6 @@ UiThemeTokens ui_theme_tokens_for_id(const char* theme_id)
 }
 
 /**
- * @brief Loads the selected theme ID from settings file.
- * @param path Settings file path.
- * @param out_theme_id Output theme ID buffer.
- * @param out_theme_id_size Buffer size.
- * @return 1 on success, 0 on failure.
- */
-int ui_theme_load_selected_id(const char* path, char* out_theme_id, size_t out_theme_id_size)
-{
-    char* text = NULL;
-    int loaded = 0;
-
-    if (!path || !out_theme_id || out_theme_id_size == 0u) {
-        return 0;
-    }
-
-    out_theme_id[0] = '\0';
-    text = file_utils_read_text_file(path);
-    if (!text) {
-        return 0;
-    }
-
-    loaded = ui_extract_json_string_value(text,
-                                          UI_THEME_CONFIG_KEY,
-                                          out_theme_id,
-                                          out_theme_id_size);
-    if (!loaded) {
-        loaded = ui_extract_json_string_value(text,
-                                              UI_THEME_COMPAT_KEY,
-                                              out_theme_id,
-                                              out_theme_id_size);
-    }
-
-    free(text);
-    return loaded && out_theme_id[0] != '\0';
-}
-
-/**
- * @brief Saves the selected theme ID to settings file.
- * @param path Settings file path.
- * @param theme_id Theme ID to save.
- * @return 1 on success, 0 on failure.
- */
-int ui_theme_save_selected_id(const char* path, const char* theme_id)
-{
-    FILE* file = NULL;
-    char* temp_path = NULL;
-    int close_result = 0;
-
-    if (!path || !theme_id) {
-        return 0;
-    }
-    if (ui_theme_index_of_id(theme_id) < 0) {
-        return 0;
-    }
-    if (!ui_theme_id_is_safe_for_json(theme_id)) {
-        return 0;
-    }
-
-    temp_path = file_utils_duplicate_path_with_suffix(path, ".tmp");
-    if (!temp_path) {
-        return 0;
-    }
-
-    file = fopen(temp_path, "wb");
-    if (!file) {
-        free(temp_path);
-        return 0;
-    }
-
-    fprintf(file, "{\n");
-    fprintf(file, "  \"%s\": \"%s\"\n", UI_THEME_CONFIG_KEY, theme_id);
-    fprintf(file, "}\n");
-    if (ferror(file)) {
-        fclose(file);
-        remove(temp_path);
-        free(temp_path);
-        return 0;
-    }
-
-    close_result = fclose(file);
-    if (close_result != 0) {
-        remove(temp_path);
-        free(temp_path);
-        return 0;
-    }
-
-    if (!file_utils_replace_file_with_temp(temp_path, path)) {
-        free(temp_path);
-        return 0;
-    }
-
-    free(temp_path);
-    return 1;
-}
-
-/**
  * @brief Applies theme tokens to Nuklear context.
  * @param ctx Nuklear context.
  * @param tokens Theme tokens to apply.
